Fix incorrect x axis date labels (#171)
# *Fix incorrect x axis date labels*

## :recycle: Current situation & Problem
- We have had several iterations on fixing the correct date formatting
for x axis in health charts, e.g. (#78 and #92). This due to how the
charts library expects the values. We display the dates in the x axis
and must supply the corresponding double value to the chart. We have
tried in the last iterations to transform the values to reduce the
distance between values in order to also reduce the spreading of the
graph.
- In the last iteration we were mapping a Date to a double of the
pattern YYYY,DD where DD was calculated as day of year / 365 rounded to
max 2 decimal places and were mapping it back to a date format the
label. Chart library on the other hand, was expecting the decimal places
to be rounded to max 2 places. Rounding was causing the issue and we
were not able to derive the initial date correctly

Adapted the approach to:
1. Given a list of measurements and their corresponding dates, pick the
oldest date
2. Map the dates as follows for the selected time range:
   - Daily: Diff in days from the oldest date
   - Weekly: Diff in weeks from the oldest date
   - Monthly: Diff in months from the oldest date
3. The diff is used ultimately as the x value, e.g. for daily time
range:

```
2025-01-01 → 0
2025-01-04 → 3
2025-01-06 → 5
2025-01-13 → 12
2025-02-14 → 44
```

4. The library supplies us back the value and expects a formatted label
text for it, e.g. 5 -> we derive the initial date as `2025-01-06` and
ultimately format it as `Jan 06`
5. Unit tests added to cover all three time ranges
6. Recording:


https://github.com/user-attachments/assets/957e625e-75a0-4362-9ded-57b1f4cc524c



## :gear: Release Notes 
*Add a bullet point list summary of the feature and possible migration
guides if this is a breaking change so this section can be added to the
release notes.*
*Include code snippets that provide examples of the feature implemented
or links to the documentation if it appends or changes the public
interface.*


## :books: Documentation
*Please ensure that you properly document any additions in conformance
to [Spezi Documentation
Guide](https://github.com/StanfordSpezi/.github/blob/main/DOCUMENTATIONGUIDE.md).*
*You can use this section to describe your solution, but we encourage
contributors to document your reasoning and changes using in-line
documentation.*


## :white_check_mark: Testing
*Please ensure that the PR meets the testing requirements set by CodeCov
and that new functionality is appropriately tested.*
*This section describes important information about the tests and why
some elements might not be testable.*


## :pencil: Code of Conduct & Contributing Guidelines 

By submitting creating this pull request, you agree to follow our [Code
of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md):
- [x] I agree to follow the [Code of
Conduct](https://github.com/StanfordSpezi/.github/blob/main/CODE_OF_CONDUCT.md)
and [Contributing
Guidelines](https://github.com/StanfordSpezi/.github/blob/main/CONTRIBUTING.md).